Working except count tokens
