+++ crates/agentgateway/src/llm/azureopenai.rs
@@ -1,9 +1,6 @@
 use agent_core::strng;
 use agent_core::strng::Strng;
-use bytes::Bytes;
 
-use crate::llm::AIError;
-use crate::llm::types::completions::typed::ChatCompletionErrorResponse;
 use crate::*;
 
 #[apply(schema!)]
