+++ crates/agentgateway/src/llm/gemini.rs
@@ -1,8 +1,6 @@
 use agent_core::strng;
 use agent_core::strng::Strng;
-use bytes::Bytes;
 
-use crate::llm::AIError;
 use crate::*;
 
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
