+++ crates/agentgateway/src/llm/openai.rs
@@ -1,8 +1,7 @@
 use agent_core::strng;
 use agent_core::strng::Strng;
-use bytes::Bytes;
 
-use crate::llm::{AIError, RouteType};
+use crate::llm::RouteType;
 use crate::*;
 
 #[apply(schema!)]
