+++ crates/agentgateway/src/llm/vertex.rs
@@ -1,6 +1,5 @@
 use agent_core::strng;
 use agent_core::strng::Strng;
-use bytes::Bytes;
 use serde_json::{Map, Value};
 
 use crate::llm::AIError;
